requirements.txt
pyrtlsdr
numpy
scipy
sounddevice
pynput